# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PyLearn classroom API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  student: "student",
} as const;

export interface UserProfile {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: UserProfileRole;
  createdAt: string;
}

export type SetRoleBodyRole =
  (typeof SetRoleBodyRole)[keyof typeof SetRoleBodyRole];

export const SetRoleBodyRole = {
  admin: "admin",
  student: "student",
} as const;

export interface SetRoleBody {
  role: SetRoleBodyRole;
}

export interface ProjectFile {
  id: number;
  userId: string;
  filename: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFileBody {
  filename: string;
  content?: string;
}

export interface UpdateFileBody {
  filename?: string;
  content?: string;
}

export interface ExecutePythonBody {
  code: string;
}

export interface ExecutionResult {
  stdout: string;
  stderr: string;
  exitCode: number;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatBody {
  message: string;
  fileContext?: string;
  filename?: string;
  conversationHistory?: ChatMessage[];
}

export interface AcceptSuggestionBody {
  fileId: number;
  newContent: string;
  lineStart?: number | null;
  lineEnd?: number | null;
}

export type HelpRequestStatus =
  (typeof HelpRequestStatus)[keyof typeof HelpRequestStatus];

export const HelpRequestStatus = {
  active: "active",
  dismissed: "dismissed",
} as const;

export interface HelpRequest {
  id: number;
  userId: string;
  /** @nullable */
  userName?: string | null;
  message: string;
  status: HelpRequestStatus;
  createdAt: string;
}

export interface CreateHelpRequestBody {
  message?: string;
}

export interface StudentInfo {
  id: string;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: string;
  isOnline: boolean;
  hasHelpRequest: boolean;
}

export type AiConfigProvider =
  (typeof AiConfigProvider)[keyof typeof AiConfigProvider];

export const AiConfigProvider = {
  openai: "openai",
  anthropic: "anthropic",
  gemini: "gemini",
} as const;

export type AiConfigMode = (typeof AiConfigMode)[keyof typeof AiConfigMode];

export const AiConfigMode = {
  off: "off",
  suggestion: "suggestion",
  agent: "agent",
} as const;

export interface AiConfig {
  id: number;
  provider: AiConfigProvider;
  mode: AiConfigMode;
  /** @nullable */
  apiKey?: string | null;
  suggestionSystemPrompt: string;
  agentSystemPrompt: string;
  offSystemPrompt: string;
}

export interface StudentLoginBody {
  /** @minLength 1 */
  displayName: string;
  /**
   * @minLength 6
   * @maxLength 6
   */
  pin: string;
}

export const StudentLoginSuccessValue = {
  success: true,
} as const;
export type StudentLoginSuccess = typeof StudentLoginSuccessValue;

export interface CreateStudentBody {
  /** @minLength 1 */
  displayName: string;
}

export interface CreatedStudentResponse {
  id: string;
  displayName: string;
  pin: string;
  createdAt: string;
}

export interface StudentAccountInfo {
  id: string;
  displayName: string;
  pin: string;
  isPaused: boolean;
  createdAt: string;
  isOnline: boolean;
  hasHelpRequest: boolean;
}

export interface ToggleStudentPauseBody {
  isPaused: boolean;
}

export type UpdateAiConfigBodyProvider =
  (typeof UpdateAiConfigBodyProvider)[keyof typeof UpdateAiConfigBodyProvider];

export const UpdateAiConfigBodyProvider = {
  openai: "openai",
  anthropic: "anthropic",
  gemini: "gemini",
} as const;

export type UpdateAiConfigBodyMode =
  (typeof UpdateAiConfigBodyMode)[keyof typeof UpdateAiConfigBodyMode];

export const UpdateAiConfigBodyMode = {
  off: "off",
  suggestion: "suggestion",
  agent: "agent",
} as const;

export interface UpdateAiConfigBody {
  provider?: UpdateAiConfigBodyProvider;
  mode?: UpdateAiConfigBodyMode;
  /** @nullable */
  apiKey?: string | null;
  suggestionSystemPrompt?: string;
  agentSystemPrompt?: string;
  offSystemPrompt?: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListFilesParams = {
  /**
   * Student ID (admin only, defaults to self)
   */
  userId?: string;
};

export type RejectSuggestion200 = {
  success?: boolean;
};

export type DeleteStudentAccount200 = {
  success: boolean;
};