# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PyLearn classroom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/workspace/{studentId}:
    get:
      tags: [admin]
      operationId: getStudentWorkspace
      summary: Get a student's workspace (files) for admin viewing
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student workspace files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectFile'
...
  /ai/chat:
    post:
      tags: [ai]
      operationId: aiChat
      summary: Send a message to AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiChatBody'
      responses:
        '200':
          description: AI response (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
  /ai/suggestion/accept:
    post:
      tags: [ai]
      operationId: acceptSuggestion
      summary: Accept an AI code suggestion and apply to file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptSuggestionBody'
      responses:
        '200':
          description: File updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFile'
...
  /admin/students:
    get:
      tags: [admin]
      operationId: listStudentAccounts
      summary: List all student accounts with PIN and status
      responses:
        '200':
          description: Student account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentAccountInfo'
    post:
      tags: [admin]
      operationId: createStudentAccount
      summary: Create a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentBody'
      responses:
        '201':
          description: Student created with plain PIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedStudentResponse'
  /admin/students/{id}:
    patch:
      tags: [admin]
      operationId: toggleStudentPause
      summary: Pause or resume a student account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleStudentPauseBody'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAccountInfo'
    delete:
      tags: [admin]
      operationId: deleteStudentAccount
      summary: Permanently delete a student account and all their data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
  /admin/ai-config:
    get:
      tags: [admin]
      operationId: getAiConfig
      summary: Get AI configuration
      responses:
        '200':
          description: AI config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiConfig'
    put:
      tags: [admin]
      operationId: updateAiConfig
      summary: Update AI configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAiConfigBody'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiConfig'
components:
...
  schemas:
...
    ProjectFile:
      type: object
      required: [id, userId, filename, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        filename:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AiChatBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
        fileContext:
          type: string
        filename:
          type: string
        conversationHistory:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
    AcceptSuggestionBody:
      type: object
      required: [fileId, newContent]
      properties:
        fileId:
          type: integer
        newContent:
          type: string
        lineStart:
          type: integer
          nullable: true
        lineEnd:
          type: integer
          nullable: true
...
    AiConfig:
      type: object
      required: [id, provider, mode, suggestionSystemPrompt, agentSystemPrompt, offSystemPrompt]
      properties:
        id:
          type: integer
        provider:
          type: string
          enum: [openai, anthropic, gemini]
        mode:
          type: string
          enum: [off, suggestion, agent]
        apiKey:
          type: [string, 'null']
        suggestionSystemPrompt:
          type: string
        agentSystemPrompt:
          type: string
        offSystemPrompt:
          type: string
...
    CreateStudentBody:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
    CreatedStudentResponse:
      type: object
      required: [id, displayName, pin, createdAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        pin:
          type: string
        createdAt:
          type: string
          format: date-time
    StudentAccountInfo:
      type: object
      required: [id, displayName, pin, isPaused, createdAt, isOnline, hasHelpRequest]
      properties:
        id:
          type: string
        displayName:
          type: string
        pin:
          type: string
        isPaused:
          type: boolean
        createdAt:
          type: string
          format: date-time
        isOnline:
          type: boolean
        hasHelpRequest:
          type: boolean
    ToggleStudentPauseBody:
      type: object
      required: [isPaused]
      properties:
        isPaused:
          type: boolean
    UpdateAiConfigBody:
      type: object
      properties:
        provider:
          type: string
          enum: [openai, anthropic, gemini]
        mode:
          type: string
          enum: [off, suggestion, agent]
        apiKey:
          type: [string, 'null']
        suggestionSystemPrompt:
          type: string
        agentSystemPrompt:
          type: string
        offSystemPrompt:
          type: string